Add subband2frequency tests

+++ src/utilities/Utils.h
@@ -18,9 +18,9 @@ namespace utils
 /**
  *
  */
-float Subband2Frequency(const int subband, const float clock=2e8f);
-float Channel2Frequency(const int subband, const int channel, const float clock=2e8f);
-float Range2Frequency(const int subband, const int start, const int end, const float clock=2e8f);
+double Subband2Frequency(const int subband, const double clock=2e8);
+double Channel2Frequency(const int subband, const int channel, const double clock=2e8);
+double Range2Frequency(const int subband, const int start, const int end, const double clock=2e8);
 
 /**
  * @brief MJDs2JD
